+++ CHANGELOG.md
@@ -2,11 +2,19 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
-## 1.4.1 - TBD
+## 1.5.0 - 2018-05-07
 
 ### Added
 
-- Nothing.
+- [#167](https://github.com/zfcampus/zf-oauth2/pull/167) adds support for PHP 7.1 and 7.2.
+
+### Changed
+
+- [#160](https://github.com/zfcampus/zf-oauth2/pull/160) alters `AuthController::tokenAction()` such that it uses the exception code from
+  a caught `ProblemExceptionInterface` instance as the ApiProblem status if it falls in the 400-600 range.
+
+- [#151](https://github.com/zfcampus/zf-oauth2/pull/151) updates `ZF\OAuth2\Provider\UserId\AuthenticationService` to allow injecting any
+  `Zend\Authentication\AuthenticationServiceInterface` implementation, not just `Zend\Authentication\AuthenticationService`.
 
 ### Deprecated
 
@@ -14,7 +22,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Removed
 
-- Nothing.
+- [#167](https://github.com/zfcampus/zf-oauth2/pull/167) removes support for HHVM.
 
 ### Fixed
 
@@ -59,6 +67,10 @@ All notable changes to this project will be documented in this file, in reverse 
 ## 1.3.3 - 2016-07-07
 
 ### Added
+
+- Nothing.
+
+### Changed
 
 - Nothing.